requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
azure-identity==1.15.0
azure-ai-ml==1.12.0
azure-cognitiveservices-language-textanalytics==5.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
